% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{execute}
\alias{execute}
\alias{execute.data.frame}
\alias{execute.NULL}
\title{Pipeline Execute function}
\usage{
execute(x, pipe, ...)

\method{execute}{data.frame}(x, pipe, ...)

\method{execute}{NULL}(x, pipe, ...)
}
\arguments{
\item{x}{input object to apply pipeline expression on.}

\item{pipe}{pipeline object}

\item{...}{additional arguments to pass to pipeline expression}
}
\value{
updated pipeline object with output
}
\description{
Execute function executes the pipeline expression on the input and stores in
output location
}
\examples{

pipe <- pipeline(expr = function(e) mean(e$mpg))
execute(mtcars, pipe)
}
