% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{flow}
\alias{flow}
\title{Pipeline Flow function}
\usage{
flow(x, pipe)
}
\arguments{
\item{x}{input object to apply pipeline expression on. Can either be
data.frame, spark dataframe or modeler object}

\item{pipe}{pipeline object}
}
\value{
data output resulting from pipeline execution
}
\description{
Flow function executes the pipeline expression on the input and returns the
data output only
}
\examples{

pipe <- pipeline(expr = function(e) mean(e$mpg))
flow(mtcars, pipe)
}
