% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline function}
\usage{
pipeline(expr = identity, desc = "", uid = random_string("pipe"))
}
\arguments{
\item{expr}{pipeline function. default is identity function}

\item{desc}{optional description input}

\item{uid}{A character string used to uniquely identify the pipeline}
}
\value{
pipeline object with expression stored
}
\description{
Creates a pipeline object to save data transformation logic. The expression
is either a named function or user defined function
}
